+++ gptables/examples/iris_quick_and_dirty.py
@@ -21,13 +21,12 @@ import gptables as gpt
 import pandas as pd
 import numpy as np
 import os
-
+from pathlib import Path
 
 ## Read data and arrange
-funcs = [np.mean, np.median]
-parent_dir = os.path.dirname(os.path.realpath(__file__))
+parent_dir = Path(__file__).parent
 
-iris_data = pd.read_csv(parent_dir + "/iris.csv")
+iris_data = pd.read_csv(parent_dir / "iris.csv")
 iris_data = iris_data.loc[:, ["class", "sepal_length", "sepal_width"]]
 
 iris_summary = iris_data.groupby("class").agg(np.mean)
@@ -55,10 +54,11 @@ iris_summary_3.insert(2, "bottom_index", ["x", "y", "z"])
 
 
 ## Write using quick and dirty method
-output_path = parent_dir + "/python_iris_quick_and_dirty.xlsx"
-gpt.quick_and_dirty_workbook(
-        filename = output_path,
-        tables = [iris_summary_0, iris_summary, iris_summary_2, iris_summary_3]
-        )
-print("Output written at: ", output_path)
+if __name__ is "__main__":
+        output_path = parent_dir / "python_iris_quick_and_dirty.xlsx"
+        gpt.quick_and_dirty_workbook(
+                filename = output_path,
+                tables = [iris_summary_0, iris_summary, iris_summary_2, iris_summary_3]
+                )
+        print("Output written at: ", output_path)